requirements.txt
streamlit
pandas
numpy
scikit-learn
Pillow
openai
tensorflow
dotenv
openai-whisper
sounddevice